// chatait-frontend-vue/src/utils/model/app/topic.ts
/*
 * Copyright 2023 Anlity <leo@leocode.net>. All rights reserved.
 * Use of this source code is governed by a AGPL v3.0 style
 * license that can be found in the LICENSE file.
 */

export interface AppConversationTopic {
  id: string
  title: string
  type: number
  tool_show_status: boolean
  edit_status: boolean
}